+++ src/Fluentify.Console.Tests/Record/SimpleWithBooleanTests/WhenWithAgeIsCalled.cs
@@ -1,0 +1,44 @@
+﻿namespace Fluentify.Console.Record.SimpleWithBooleanTests;
+
+public sealed class WhenWithAgeIsCalled
+{
+    [Fact]
+    public void GivenNullSubjectThenArgumentNullExceptionIsThrown()
+    {
+        // Arrange
+        SimpleWithBoolean? subject = default;
+
+        // Act
+        Func<SimpleWithBoolean> act = () => subject!.WithAge(1);
+
+        // Assert
+        _ = act.Should().Throw<ArgumentNullException>()
+            .WithParameterName(nameof(subject));
+    }
+
+    [Theory]
+    [InlineData(1)]
+    [InlineData(-1)]
+    [InlineData(0)]
+    [InlineData(int.MinValue)]
+    [InlineData(int.MaxValue)]
+    public void GivenAnAgeThenTheValueIsApplied(int age)
+    {
+        // Arrange
+        var original = new SimpleWithBoolean
+        {
+            Age = Random.Shared.Next(),
+            IsRetired = true,
+            Name = "Avery Brooks",
+        };
+
+        // Act
+        SimpleWithBoolean actual = original.WithAge(age);
+
+        // Assert
+        _ = actual.Should().NotBeSameAs(original);
+        _ = actual.Age.Should().Be(age);
+        _ = actual.IsRetired.Should().Be(original.IsRetired);
+        _ = actual.Name.Should().BeEquivalentTo(original.Name);
+    }
+}